backend/dashboardtemplatev1/requirements.txt
asgiref==3.7.2
build==1.0.3
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
importlib-metadata==6.8.0
packaging==23.2
pip-tools==7.3.0
psycopg==3.1.12
psycopg2==2.9.9
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
zipp==3.17.0
